Clutch calibration confirmed working
 - Also fixed EEPROM bug with 32U4
- Compile fixes for 32U4
- Removed compile errors

+++ DahlDesignDDC/03_EEPROM.ino
@@ -30,7 +30,7 @@ bool DDS_b_init = true;
 bool DDS_s_init = true;
 int oldBitePoint;
 int oldLED;
-int oldBrake;
-int oldThrottle;
+float oldBrake;
+float oldThrottle;
 int oldPreset;
 #endif
